+++ src/main/java/com/gamingmesh/jobs/container/JobsPlayer.java
@@ -440,7 +440,7 @@ public class JobsPlayer {
 //	synchronized (saveLock) {
 	if (!isInJob(job)) {
 	    int level = 1;
-	    int exp = 0;
+	    double exp = 0;
 
 	    JobProgression archived = getArchivedJobProgression(job);
 	    if (archived != null) {
@@ -486,9 +486,10 @@ public class JobsPlayer {
 	return level;
     }
 
-    public int getExpAfterRejoin(JobProgression jp, int level) {
+    public double getExpAfterRejoin(JobProgression jp, int level) {
 	if (jp == null)
 	    return 1;
+
 	Integer max = jp.getMaxExperience(level);
 	Double exp = jp.getExperience();
 	if (exp > max)
@@ -503,7 +504,8 @@ public class JobsPlayer {
 	    } else
 		exp = (exp - (exp * (Jobs.getGCManager().levelLossPercentage / 100.0)));
 	}
-	return exp.intValue();
+
+	return exp.doubleValue();
     }
 
     /**
